Fix updateButtons in JoystickLib
Before this fix, PreviousButtonState's members were being set to CurrentButtonState's members values AFTER CurrentButtonState was updated, which means that previous is always going to be equal to current. This is fixed with this commit, which changes the order of assignment. Additionally, rather than copying the array instance, it sets PreviousButtonState's members, so that it remains a unique array. Arrays are objects in python. Maybe you meant to use the array copy operator? This would look like 
PreviousButtonState = CurrentButtonState [:]
Additionally, you should always initialize a variable before using it. Python can't recognize this in an OOP context, but it's important nonetheless.
In any case, this code should work now. 

-Liam Taylor, Head programmer for 2605 from 2013-2015